Reorganize the wasm node by moving files around
