+++ wasm-node/javascript/src/internals/base64.ts
