+++ wasm-node/javascript/src/internals/buffer.ts
