+++ wasm-node/javascript/src/internals/local-instance.ts
