+++ wasm-node/javascript/src/module-deno.ts
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import { default as wasmBase64 } from './module/autogen/wasm.js';
+import { default as wasmBase64 } from './internals/module/wasm.js';
 
 /**
  * Compiles and returns the smoldot WebAssembly binary.
